src/main/java/Algorithm/arr/MoveZeroes.java
package Algorithm.arr;

public class MoveZeroes {
    public void moveZeroes(int[] nums) {
        int i = 0;
        int j = 1;
        while (j < nums.length) {
            if (nums[i] == 0 && nums[j] != 0) {
                swap(nums, i++, j++);
            } else if (nums[i] == 0 && nums[j] == 0) {
                j++;
            } else {
                i++;
                j++;
            }
        }
    }

    private void swap(int[] nums, int i, int j) {
        int temp = nums[i];
        nums[i] = nums[j];
        nums[j] = temp;
    }
}
